requirements.txt
lxml==4.5.1
requests==2.23.0
pyyaml==5.3.1
python-slugify==4.0.0